requirements.txt
streamlit
google-generativeai
python-dotenv
requests
google_places 
reportlab
streamlit_extras
streamlit_folium
geopy
folium
